\name{TP53.test1.cancer}
\alias{TP53.test1.cancer}
\docType{data}
\title{
Cancer information data
}
\description{
Cancer information data used in function Famdenovo
}
\usage{TP53.test1.cancer}
\format{
  A data frame with 3 variables.
  \describe{
    \item{\code{id}}{Individual id}
    \item{\code{cancer.type}}{Cancer type. See \link{LFSpro.cancer.type} for details}
    \item{\code{diag.age}}{The age when the individual was diagnosed with cancer}
  }
}
\examples{
data(TP53.test1.cancer)
TP53.test1.cancer
}
\keyword{datasets}
