\name{TP53.test2.counselees}
\alias{TP53.test2.counselees}
\docType{data}
\title{
The information of De novo or familial status (MA Anderson Cancer Center)
}
\description{
Information of of De novo or familial status used in function Famdenovo
}
\usage{TP53.test2.counselees}
\format{
  A data frame containing 2 variables.
  \describe{
    \item{\code{id}}{individual id}
    \item{\code{con.state}}{status of de novo or familiar of an individual person.}
  }
}
\examples{
data(TP53.test2.counselees)
TP53.test2.counselees
}
\keyword{datasets}
